+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* [Disclaimer](ABOUT.md), regarding the objectives and limitations of this project.
+
 ## [0.7.0] - `2019-10-19`
 
 ### Added