+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * [Disclaimer](ABOUT.md), regarding the objectives and limitations of this project.
 
+### Changed
+* Bumped [RuboCop](https://github.com/rubocop-hq/rubocop) to `v.0.76.0`.
+
 ## [0.7.0] - `2019-10-19`
 
 ### Added